Refactor the viper builder to separate file
